CogVM source as per VMMaker.oscog-eem.2885
Fix the regression in db05a53a9f41dc87aafed07cc83d1ff22b46d42c/
VMMaker.oscog-eem.2876. Since pinning is potentially a become operation
followForwardingPointersInStackZone: is invoked, and that crashes when
stackPage is 0, as is it is when snapshotting.  So the fix is to return to not
pinning in postGCUpdateDisplayBits if snapshotting.

Slang:
Fix a regression with inlining isClassOfNonImm:equalTo:compactClassIndex
  Not quite sure why but an argument assignment for the unused classOop
argument was being generated when it shouldn't be.

Cogit: save several lines by inlining wrappers around
genTrampolineFor:called:numArgs:arg:arg:arg:arg:regsToSave:pushLinkReg:resultReg:appendOpcodes:.  N.B. This breaks translation of the SimpleStackBasedCogit but
we don't care.  It can be simulation only.