CogVM source as per  VMMaker.oscog-eem.2943
Resolve the conflict between sqAssert.h's attempt to export warning and error to
external plugins with the crude redefinition of the EXPORT macro on generating
plugins. So so by moving the redefinition of EXPORT after a plugin's include
files. The issue here is Windows' requirement to use declspec(dll_import) &
declspec(dll_import) for sharing between exe and dll.