Play the features.h dance
Since configure takes care that `_GNU_SOURCE` is there
whenever supported, _not_ having that means we cannot reasonably
ensure that we can access reg names, as defined via `__USE_GNU`.

The `features.h` makes sure that `__USE_GNU` is set if and only if
`_GNU_SOURCE` is set. So Bailing when not having it is the right choice.