Fix type of halftone array for 64-bit targets
The halftone array is accessed using a hard-coded multiplier of 4 bytes,
therefore the type of each element needs to be 32 bit on every platform.
`sqInt` is not appropriate for this use, since it is a 64-bit type on
64-bit platforms. Rather than unilaterally introduce C99 stdint types,
use `unsigned int` since this wil be 32-bit on both current fast path
binary targets.