CogVM source as per  VMMaker.oscog-eem.3009
Spur: Fix a bad bug with computing accessor depths. Exported primitives in the
StackInterpreter and subclasses had their accessor depths computed after
inlining, which is too late.

Add VM parameter 76 to answer the minimumUnusedHeadroom on a stack page.
This is useful in checking that a FastCPrimitiveFlag primitive has not
consumed too much stack.