CogVM source as per VMMaker.oscog-eem.3011
Cogit: Get the stack adjustment after primitive success right for
FastCPrimitiveFlag primitives, with or without the FastCPrimitiveAlignForFloatsFlag.

Support a FastCPrimitiveAlignForFloatsFlag to accompany FastCPrimitiveFlag.
Use the facility in primitiveFFI{Double,Float}At[Put]. On win32 clang emits
instructions that insist on at least a 16 byte aligned stack.

Hence identify fast machine code primitives with the FastCPrimitiveUseCABIFlag
flag, refactor compileOnStackExternalPrimitive: to
compileOnStackExternalPrimitive:flags:, extend the set of PrimCall flags, etc.

Remember to set the native stack pointer for a FastCPrimitive on ARMv8.
This is almost certainly the cause of current Cog crashes on ARMv8.