CogVM source as per Name: VMMaker.oscog-eem.3022
Cogit: FastCPrimitiveCall.
Get the receiver and restore retpc correctly on CISCs.
Save a few bytes setting argumentCount after zeroing primFailCode in C primitive invocation.

Save and restore LinkReg and/or SPReg around the primitive and checkForAndFollowForwardedPrimitiveState calls.

Spur CoInterpreter: Improve checkForAndFollowForwardedPrimitiveState. Allow the method's first literal (if an FFI or named primitive) to be forwarded, unfollowing it *before* accessorDepth is fetched.  Follow it to the relevant depth.

Cog: use symbolic constants for Spur primitive metadata manipulation.
Refactor linkExternalCall:ifFail: into linkExternalCall:errInto: so it doesn't have to be inlined.
Fix mapToBytecodePCIfActivationOfExternalMethod: to asvoid an assert fail for an interpreter frame activation of an external method.
Use faster length accessors (numSlotsOf: numBytesOfBytes:) in a few places.

Reorder the storing of newMethod and the setting of stack pointers in compileOnStackExternalPrimitive:flags:, as well as making an invariant on VarBase clear.

Merge Ben Avison's minor changes that help support the improvements he made for the ARM64 bitblt

Have the CroquetPlugin primitives use the FastCPrimitiveFlag FastCPrimitiveAlignForFloatsFlag flags.  Use the methodReturnFoo: protocol.

Use symbolic constants for the Spur primitive metadata manipulation.

Don't pull in ClosedVMMaker when building a trunk VMMaker image.  ClosedVMMaker
is history.