CogVM source as per VMMaker.oscog-eem.3024
Cog: add a trace flag to turn on logging of FastCCall primitives.
Mark a few of the flag methods in the Cogit as <inline: #always>
and make compileOnStackExternalPrimitive:flags: a Spur-only option
(it is anyway, so the code shouldn't be incuded in V3 Cogits).
Remove the inline marker from a couple of api methods.

Plugins regenerated as per VMMaker.oscog-eem.3014.