CogVM source as per VMMaker.oscog-eem.3041
Cogit: nuke the one instance of a C ABI interpreter primitive,
mcprimHashMultiply:.  It now has a machien code implementation and so the
interpreter version and its support machinery are now dead weight.  Replace
this by setting the FastCPrimitive flag for CompiledCode>>#objectAt:put:,
primitiveObjectAtPut, which yields about a 4% speedup for e.g. Morph compileAll.

ThreadedFFIPlugin:

Complements FFI-Pools-mt34 and FFI-Kernel-mt.207.

Current version of the FFI plugin is 1. Changes to the atomic-type codes or known classes in special-objects array should bump that version number so that image code can adapt.