CogVM source as per VMMaker.oscog-eem.3061
SpurMemoryManager:
Fix a regression in image segment loading introduced by VMMaker.oscog-eem.2819.
The old image segment array shortening code created a header ambiguous with a
slimbridge if the array was in new space.  In this case we also have to shorten
the array's normal slot count so its header is taken as a slimbridge.

newSpaceLimit is equal to oldSpaceStart and therefore surplus to requirements.