CogVM source as per VMMaker.oscog-eem.3073
Slang: add mechanism for an internal plugin to include additional exports, hence allowing the functions in sqFFITestFunctions.c to be exported from SqueakFFIPrims when compiled as an internal plugin.

SqueakFFIPrims: add sqFFITestFuncExports.h which exports the test functions when
SqueakFFIPrims is compiled as an internal plugin.  Move the declarations and
types to sqFFI.h.

Don't compile/export test functions in IA32ABI unless ALIEN_TEST_FUNCTIONS is
defined on the command line, hence resolving conflict with SqueakFFIPrims'
test functions.

Add a hack to sqNamedPrims.c to allow it to answer an internal module from
ioLoadModuleOfLength.  Note that search through internal exports in
findInternalFunctionIn is exhaustive (!!).  This cannot stand ;-)

Fix the Mac makefiles to remove the dependence on the dep file,
as this was forcing recompilation all the time.