CogVM source as per VMMaker.oscog-eem.3074
CoInterpreter:
Fix bad bug in justActivateNewMethod:.  If newMethod was cogged and the method
must be activated as an interpreter method (ie primitiveFailForFFIException:at:)then the old code did not access the true methodHeader, causing a crash.

Eliminate unnecessary instructionPointer casts; it is unsigned.

On Mac reorganize setting signals to eliminate unnecessary call in stack VMs
and add a secret option to cause the VM to wait for a debugger to attach on
receiving a sigsegv (a limitation of lldb on MACH). This allowed the above issue
to be debugged.