CogVM source as per VMMaker.oscog-eem.3101
Interpreter/Simulator:
Revise and refactor frame printing to allow the simulators to prefix frame
addresses with an indication of the various frame/stack pointers (variables,
processor registers).  Extract the printing of the frame address to
printFrameAddress: there-by eliminating quite a few simulator-specific versions
of the frame field printers.

Use %p to print pointers, using extensions to the Printf package, '%P' to print
with the 16r prefix, and '%WP' to print with 16r prefix in a width dependent on
the word size.

Realise that %Ns prints a left-padded string in a width of N characters to
eliminbate the torturous %.*s%s forms using spaces.

Since %p always requires a pointer, cast the arguments to pringHex: et al to
void pointers, and hence get rid of printHexPtr:.

Fix a slip in ObjectMemory>>#printNonPointerDataOf:on:.

Interpreter printing: must use %s to print object strings since they may
contain %'s.

Spur Slang: fix a bug in computing accessor depth chains.  If a primitive
contains different assignments to the same variable (e.g. a primitive which has
different forms for var args such as PSharePlugin>>primitivePostKeyboardEvent
as of 10/26/2021) then the simple test would consider an assignment to the
same variable as an extension of the chain.