CogVM source as per VMMaker.oscog-eem.3108
Fix a crash in recursive invocations of primitive 118,
[receiver:]tryPrimitive:withArgs:.  This is a bit of a hack; nested primitive
failure will not be unwound correctly in tgh eouter invocation.
We may have to revisit this, but at least the simple case of
	6 tryPrimitive: 118 withArgs: {9. {}}
no longer crashes the VM.

CogProcessorAlien plugins:
Implement a memory offset to allow the simulatorw to implement
a null pointer trap by omitting the first word in simulated memory.