CogVM source as per VMMaker.oscog-eem.3109
Cogit: add machine code versions of primitiveSlotAt[Put] (used for
instVarAt:[put:] in Spur) for Spur64.  This is mainly to reduce noise in
Virtend crash log primitive traces, but is a lot faster.

Spur: Fix a bug in full GC with non-writable code zone. The scavenge that comes
immediately before the mark phase in a full GC needs to be done with
GCModeNewSpace to ensure the Cogit's GC routines are invoked correctly.