CogVM source as per Name: VMMaker.oscog-eem.3127
Revise primitiveSuspend to no longer just remove a process from its
condition variable list if waiting on a condition variable. Instead
also back up the process one bytecode so that the blocking send will
be retried if and when the process is resumed.

Distinguish between a process being runnable but not active from a
blocked process by the class of myList.  If myList is LinkedList the
process is runnable.  Hence keep track of the class of LinkedList.

Refactor convertToInterpreterFrame: to convertFrame:toInterpreterFrame:
to facilitate implementation in the Cogit (a send byetcode is not a
suspension point in machien code, hence a machien code caller of a
blocking primtiive must be converted to an interpreter frame).

Fix formatting bug in longPrintOop:.

Cog: several places can use startPCOfMethodHeader: instead of startPCOfMethod:

convertToInterpreterFrame: needs to assert validInstructionPointer:..., not
just evaluate it.

Interpreter: better to test process having a context first in printAllStacks.