CogVM source as per VMMaker.oscog-eem.3132
Interpreter:
Have primitiveSuspend fail for unresumable processes.
Beef up removeProcess:fromList: to follow forwarders in the list, allowing
become on suspended processes (this weas definitely a bug), and have it fail
if it encounters a link that doesn't look like a process.
Rename backupProcess:toBlockingSendTo: to backupContext:toBlockingSendTo:
because that's what it does and it saves an access

Slang: provide an x ordinate in the accessorDepthTable to make it easier to
lookup accessor depths