CogVM source as per VMMaker.oscog-eem.3142
Change the revised primitiveSuspend to behave clkoser to the original, always
answering the list the receiver was on if bnlocked.  make the newer variant
available as a named primitive (primitiveSuspendV2). AFAICT using
primitiveSuspendV2 breaks lots of things (SqueakSSL tests hang, Virtend login
hangs).  So while the idea is right, in practice answering nil if the receiver
is suspended while waiting on a condition variable breaks more things than it fixes.