CogVM source as per Name: VMMaker.oscog-eem.3153
Primitive suspend: revert the semantics of #88 to Andreas' revision in the early
2000's.  #88 removes a process from a condition variable, allowing subsequently
resumed processes to get past their condition variable.  This is a bug, but
there are images (noably Qwaq/Teleplace/Virtend) which depend on this behaviour.
Provide #568 (primitiveSuspendBackingUpV1) which backs up a process waiting on a
condition variable to the send that invoked the wait primitive, and which
answers that list. Provide #578 (primitiveSuspendBackingUpV2) which backs up a
process waiting on a condition variable to the send that invoked the wait
primitive, but in which case answers nil. The presence of the three primitives
is indicated by bit 5 of the cogVMFeatureFlags.

DeflatePlugin: the update primitives can run on the Smalltalk stack.

Add the FileDialogPlugin from Qwaq/Teleplace/Virtend.