CogVM source as per VMMaker.oscog-eem.3160
CoInterpreter: fix a serious bug with post-become processing in Spur where the
instruction pointer could be updated incorrectly if it was in machine code
(CoInterpreter>>followForwardingPointersInStackZone:).

Add shared pool variables to cointerp.h so that phase errors are caught.
Rename PrimErrTableIndex to PrimitiveErrorTableIndex so it is not confused with the PrimErrXXX values.

Update a few primitives that use PrimCallMayEndureCodeCompaction to also supply the required PrimCallNeedsNewMethod.