For macOS platform, fixes the sporadic segfaults...
...that started to happen after the recent changes to both -metal and -core-graphics backends, which entailed a slight change in what is actually in the view's event loop. That is, we added "redraw" events to work around vsync limitations. See uses of setNeedsDisplayInRect:.

The solution was, apparently, that we must not call the ancient Carbon-style pumpRunLoop() but always do the modern pumping via pumpRunLoopEventSendAndSignal().

Still, one of the remaining issues is that, for example, Squeak's #disableDeferredUpdates does not work for MVC-style display, which relies on the #copyBits side effect of directly pushing the contents to the screen. I will think of something such as pumping events from within the image.