SqueakFFIPrims source as per VMMaker.oscog-mt.3181
In the FFI plugin, fixes a regression in int64_t support on 32-bit builds, which was due to a mix-up between int64_t and intptr_t.

Note that the current design only supports up to 64-bit integer types. If we want to support, for example, 128-bit integer types on a 64-bit machine, we would have to think about a replacement for 'usqLong'. Or even better, we find a way to specialize #ffiArgByValue:in: in subclasses of ThreadedFFIPlugin to use the stack more efficiently.