[SqueakSSL/unix/openssl] Fix for OpenSSL3
OpenSSL 3 deprecated SSL_get_peer_certificate, made it a macro and an
provides SSL_get{0,1}_peer_certificate instead.

Why even more complicating that already messed API is beyond me.
Not promising ABI stability is well known for OpenSSL but, why the HECK
not simply providing an Alias Symbol instead of UNNECESSARILY ripping
the symbol out of that innocent object file?

Also, we now need a load-time disambiguation of which symbol to use.

(if you find a rant in this message, you can keep it)