Fix suspiscious check in Win32 socket plugin
>/usr/x86_64-w64-mingw32/sys-root/mingw/include/WinSock2.h:1029:88: note: passing argument to parameter 'optval' here
>  WINSOCK_API_LINKAGE int WSAAPI setsockopt(SOCKET s,int level,int optname,const char *optval,int optlen);
>                                                                                       ^
>../../../platforms/win32/plugins/SocketPlugin/sqWin32NewNet.c:1758:7: warning: logical not is only applied to the left hand side of this comparison [-Wlogical-not-parentheses]
>  if (!opt == 0
>      ^    ~~

Since opt is a pointer, !opt means (pointer is NULL)<br>
Then (!opt) == 0, means (pointer is NULL) is false, hence (pointer is not NULL).

So we would barf if pointer is not NULL?<br>
But on next line (reached if pointer IS NULL), we would dereference the pointer...<br>
Err, non-sensical.

I guess the intention was to BARF if pointer is NULL.