CogVM source as per VMMaker.oscog-eem.3264
Fix regression in emission of primitive metadata in plugins that I think was
introduced in 3222 or there-abouts.  The bug is when vmClass is set then a
selector must begin with prim to have metadata generated. I *think* I assigned
vmClass to compute accessor depth correctly for plugin primitives around this
time, but I'm not sure.

The fix is to have VMPluginCodeGenerator override methodNeedsMetadata: and not
insist that the selector begin with 'prim'.