CogVM source as per VMMaker.oscog-eem.3268
Fix several sends of [positive|signed]MachineIntegerFor: which is implemented
in InterpreterPlugin, not InterpreterProxy.

Refactor call logging in the ThreadedFFIPlugin to get as much of the logging
code out of the critical path as possible.

Fix issues with spur primitive metadata in plugin primitives.
stackFloatValue: stackStringValue: et al look inside the object on the stack.
stackValue: stackIntegerValue: et al do not. So extend
StackInterpreter class>>isObjectAccessor: to include those stack accessors
that look inside the object on the stack.  This fixes the metadata calculations
for things like B3DAcceleratorPlugin>>primitiveSetLights,
FileDialogPlugin>>primitiveFileDialogAddFilter, etc.