Use appropriate function to destroy the currentCursor in Win32
The currentCursor is either created by CreateCursor() in ioSetCursorWithMask() or by CreateIconIndirect() in ioSetCursorARGB()
Thus it must be destroyed either by DestroyCursor() or by DestroyIcon()

A new global variable currentCursorIsIcon memoize how the currentCursor was created and thus how it must be destroyed

if ever the creation of cursor fails, restore the old cursor rather than letting it be null.
Thus, only destroy the old cursor if the creation of new cursor succeeds.