CogVM source as per VMMaker.oscog-eem.3280
Add a primitive that computes the transitive closure of objects accessible
from an array of roots.  This is the collection half of the storeImageSegment
primtiive.  This can be used either to debug failures in the storeImageSegment
primitive, and may eventually be used to implement it in a more general way.

Thanks to Igor Stasenko for suggesting this decomposition a few years ago.