CogVM source as per VMMaker.oscog-eem.3287
Fix become on compiled methods (for Spur only).  Tim Rowledge's recent attempt
to inboard sources uncovered flaws in become with compiled methods on some
jitted platforms (specifically ARMv8 linux, which has a dual mapped code zone).

The main changes in this are to
a) never do in-place become on compiled methods on the jit
b) divorce machine code frames referring to cog methods involved in a become
c) refactor to reduce duplication and overrioding in spurPostBecomeAction:
d) fix bugs in mapping the instructionPointer when method and/or newMethod are
   becommed
e) fix bugs in exception handler search that was not following forwarding
   pointers and could hence possibly be confused (miss finding a handler)
   by a become

Hide access to CogBlockMethod cmType behind accessors, so as to allow CMMethod
to be extended to address the above issue with become in an elegant way.