CogVM source as per VMMaker.oscog-eem.3297/Balloon3D-Plugins-eem.20
Include a small suite of VM benchmark primitives to measure the time taken by
operations that are too short-lived to want to time them (because timing can
have significant overhead).

Add an abstraction for identifying full block methods and use it in asserts.

Refactor followForwardingPointersInStackZone: to followForwardingPointersInStackZone
since it makes no use of the becomeEffectFlags argument.

Make sure all uses of cloneObject: in plugins test the result appropriately.

Make sure the head frame pointers are written back when cloning a context in
primitiveClone.

Simplify the TempVectReadBarrier testing in Stack/CoInterpreter, adding one
method and deleting four, and testing TempVectReadBarrier in a single method
only (albeit one that is overridden in CoInterpreter).

Given that better forwarding machinery was written over time, use a slightly
better form for dealing with potential forwarding in the pushFullClosure bytecode,
and neaten StackInterpreter>>pushLiteralVariable:/storeLiteralVariable:withValue:
to use it.