CogVM source as per VMMaker.oscog-eem.3302
StackInterpreter: fix a bad bug introduced by VMMaker.oscog-eem.3287.
VMMaker.oscog-eem.3287 fixed relocating the instructionPointer by remapping
either relative to method or newMethod, depending on which one contains
instructionPointer.  But if not done carefully, it is possible to relocate
the instructionPointer twice.

Spur: add a halt to a objectsReachableFromRoots: failure case, agreeing with
the convention elsewhere in the ImageSegment primitives.

Cogit: inline StackToRegisterMappingCogit>>genPushLiteral: which simply calls
ssPushConstant: