CogVM source as per VMMaker.oscog-eem.3303
Do a more precise implementation of VMMaker.oscog-eem.3302.  In the
StackInterpreter instrucftionPointer is never relative to newMethod;
this only occurs in the Cogit with machine code primitives invoking
C primitive implementations.  Further, the basic fix to mapVMRegisters
to avoid the hazard of updating instructionPointer twice also needs
to be applied to followMethodNewMethodAndInstructionPointer.