CogVM source as per VMMaker.oscog-eem.3316
Bow to the inevitable and go with as traditional platforms/Cross style include
file for the ClipboardExtendedPlugin (int is not wide enough to hold an oop on
64-bits).

Harden the fetch functions on iOS and unix to check for out-of-memory.

The plugin now seems to work on unix but much work needs to be done to
platforms/unix/vm-display-x11/sqUnixX11.c because playing with e.g. firefox
shows that copying HTML text has no effect on the results passed up from
clipboardGetTypeNames et al.