Sligthly cleanup joystick plugin on unix
note that the event code header itself is unneccessary, it has always
been included by input.h itself. So, if it is there, it has been
included regardless, and if it it not there (old linux), it will
complain. So remove that.
Also, include sq.h as customary for plugins and thus lose a few
common headers