CogVM source as per VMMaker.oscog-eem.3349
Integrate VMMaker.threaded-LM.3339, 3341, 3342 & 3343.

Make vmThread>>#state an atomic_int

Allow -1 to be set as vmOwner.
This is currently used by the heartbeat thread.
In future this will be used by the ownVMFromUnidentifiedThread function.

This combines well with the additional change that makes disownVM return a CogVMThread* disguised as void*.

We can then ensure that the threads variable is only ever accessed by the vm owner.

Fix multiple issues regarding thread safety

1. vmOwner -1 can be used to lock the VM to a thread that is not a classical "VM Thread" (e.g. the heartbeat thread).
2. maxWaitingPriority is now an atomic.

Fix thread switch accidentally disowning the VM which causes an incorrect preemption.

Allow thread switching during the SqueakSSL primitiveConnect.

Rename ownerIndex to threadAffinity.

And update revertIfEssentiallyUnchanged to cope with the CryptographyPlugins.