CogVM source as per VMMaker.oscog-eem.3353
VMMaker side of embedded images.
Provide a function checkImageHeaderFromBytes:AndSize: that can be used
by the platform to identify likely embedded image candidates.

Answer vmParameterAt: 50 with two boolean flags (encoded in an integer)
of whether the image loaded with swizzling or not, and whether the image
is embedded or not.

Add a default definition for sqImageFileIsEmbedded