CogVM source as per VMMaker.oscog-eem.3419
Now that <inline: #always> is less affected by a less capricious inliner use
it to shed code from the generated BitBlt plugin, and the cogits.

Slang:
Change the record of inlining phase so that it is known when inlining has yet
to start, has started, and has finished, instead of only having begun or not.
Refactor the doFinalValidationsAndTransformations and make sure it is done
along all inlining code paths. Pull the check for required inlinability out of
inlineableFunctionCall:in: & inlineableSend:in: and into
doFinalValidationsAndTransformationsIn: where it is only performed once
inlining has finished.

Fix isFunctionalIn:. Fix the bug in hasReturnOnlyAtEnd that failed to find
returns with the blocks of a returning if that didn't come at the end of
those blocks.
Fix TStmtListNode>>#copyWithoutReturn to handle trailing returning if's
correctly.
Fix TStmtListNode>>#replaceNodesIn: to flatten inlined statement lists.