TARGET_OS_IS_IPHONE is an oddball
First, it was only ever defined in John's projectBuilder/Xcode files,
and THEN even withot a value, thus NEVER true in #if.
Also, it is elsewhere only use with #ifdef/#if defined.

And Then, TARGET_OS_* is a probably Next/OpenStep descendent thing,
(see your local TargetContitionals.h),
BUT these are NEVER TARGET_OS_IS_... but only TARGET_OS_...

We had warngins for that for a long time and Clang 18 is having noe of
that anyomre.

This changes the #if to an #if defined.
PROBALBY All references to TARGET_OS_IS_IPHONE should be nixed.