CogVM source as per VMMaker.oscog-eem.3473
ThreadedFFIPlugin

Add support for varargs ('...' punctuator) that allows ThreadedARM64AppleFFIPlugin
to funciton correctly on macOS Apple Silicon. The issue here is that all optional
parameters are passed on the stack (which implies the callee does not have to save
its register arguments).

Annoyingly Apple doesn't document this clearly, but such is life.

Cache nilObj since it doesn't change and it is used often enough in argument
marshalling.

Reorder and retype some variables in the callout struct for a slightly more
compact struct.