CogVM source as per VMMaker.oscog-eem.3482
Spur:
Refactor ephemeron firing so that fireEphemeron: is a method on
SpurMemoryManager, not StackInterpreter.

Make sure that queued/fired ephemerons are made to look unqueued/unfired
in a snapshot (since snapshots don't save the mourn queue).

Fix an invalid assert in markAllUnscannedEphemerons. It could indeed be that
there are no unscanned ephemerons on invoking markAllUnscannedEphemerons.

Fix a slip in the validImage leak checker which produced a spurious cr
for each bad object when run without -verbose.

Add ephemeron and weakling printing routines.

Interpreter:
Given that MULTIPLEBYTECODESETS is translation time only, don't use
cppIf: MULTIPLEBYTECODESETS, use MULTIPLEBYTECODESETS ifTrue: ...

Fix the Gnuifier for this change which was including an

Fix primitiveBytecodeSetsAvailable for the case of
MUTLIPLEBYTECODESETS being false.

Have primitiveGetenv use stackEphemeralStringValue: and methodReturnString:.