CogVM source as per VMMaker.oscog-eem.3491
Fix a bug in managing forwarders in the remembered set dating
from mid September '13.  Forwarders in the remembered table do
not need to be handled specially. Btu they are yet another case in
scavengeReferentsOf:. Rename the subroutines of scavengeReferentsOf:
to allow for consistent naming and add scavengeReferentOfForwarder:.

Add checking the objStacks and remembered set to printReferencesTo: