[unix/make/plugins] reduce export-dynamic
export-dynamic is a linker flag and as of 2025 can upset clang in
compile mode. so drop it _except_ for the libtool invocation, where
it is apparently correct.

NB: IF it was for plain compile-link-cycles, the correct form would be
`-Wl,-export-dynamic`, but as we are using libtool in this case, that
shall not be necessary